+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project are documented here.
 
+## Unreleased
+
+### Added
+
+- English and Simplified Chinese localization for theme settings, following the DSH page language.
+
 ## [0.1.0] - 2026-08-26
 
 ### Added